Updates to make CX more compatible with Linux. The main thing was renaming CX::Instances::Display to CX::Instances::Disp to deal with the obnoxious Display struct in the X11 header. All examples brought up to date with new name. Setup of OpenGL was improved to be more compatible with older hardware/Linux drivers. A number of the changes in this commit are actually line ending disagreements between Windows and Linux, not legitimate changes.

+++ src/CX_SoundBuffer.cpp
@@ -67,10 +67,10 @@ bool CX_SoundBuffer::loadFile (string fileName) {
 			FMOD_RESULT lockResult = FMOD_Sound_Lock( fmSound, 0, samplesToRead, &ptr1, &ptr2, &length1, &length2 );
 
 			if (lockResult == FMOD_OK) {
-				unsigned int totalSamples = length1 * channels; //The documentation says that length1 is in bytes, but it seems to be wrong. 
+				unsigned int totalSamples = length1 * channels; //The documentation says that length1 is in bytes, but it seems to be wrong.
 				//It is in samples for everything I've found (which have been entirely 16-bit PCM).
 				_soundData.resize( totalSamples );
-				
+
 				for (unsigned int i = 0; i < totalSamples; i++) {
 					_soundData[i] = ((float)(((int16_t*)ptr1)[i]))/32768;
 				}
@@ -204,7 +204,7 @@ bool CX_SoundBuffer::addSound(CX_SoundBuffer nsb, CX_Millis timeOffset) {
 
 	//Get the new data that will be merged.
 	vector<float> &newData = nsb.getRawDataReference();
-	
+
 	//If this sound isn't long enough to hold the new data, resize it to fit.
 	if (insertionSample + newData.size() > this->_soundData.size()) {
 		_soundData.resize( insertionSample + newData.size(), 0 ); //When resizing, set any new elements to silence (i.e. 0).
@@ -221,10 +221,10 @@ bool CX_SoundBuffer::addSound(CX_SoundBuffer nsb, CX_Millis timeOffset) {
 	return true;
 }
 
-/*! Set the contents of the sound buffer from a vector of float data. 
-\param data A vector of sound samples. These values should go from -1 to 1. This requirement is not checked for. 
-If there is more than once channel of data, the data must be interleaved. This means that if, for example, 
-there are two channels, the ordering of the samples is 12121212... where 1 represents a sample for channel 
+/*! Set the contents of the sound buffer from a vector of float data.
+\param data A vector of sound samples. These values should go from -1 to 1. This requirement is not checked for.
+If there is more than once channel of data, the data must be interleaved. This means that if, for example,
+there are two channels, the ordering of the samples is 12121212... where 1 represents a sample for channel
 1 and 2 represents a sample for channel 2. This requirement is not checked for. The number of samples in this
 vector must be evenly divisible by the number of channels set with the `channels` argument, which is checked for!
 \param channels The number of channels worth of data that is stored in `data`.
@@ -250,7 +250,7 @@ bool CX_SoundBuffer::setFromVector(const std::vector<float>& data, int channels,
 \param channel The channel to set the data for. If greater than any existing channel, new channels will be created
 so that the number of stored channels is equal to `channel + 1`. If you don't want a bunch of new empty channels, make sure
 you don't use a large channel number.
-\param data A vector of sound samples. These values must be in the interval [-1, 1], which is not checked for. 
+\param data A vector of sound samples. These values must be in the interval [-1, 1], which is not checked for.
 See CX::Util::clamp() for one method of making sure your data are in the correct range.
 If the other channels in the CX_SoundBuffer are longer than `data`, `data` will be extended with zeroes.
 If the other channels in the CX_SoundBuffer are shorter than `data`, those channels will be extended with zeroes.
@@ -330,7 +330,7 @@ with an absolute value greater than or equal to tolerance is removed from the so
 */
 void CX_SoundBuffer::stripLeadingSilence (float tolerance) {
 	for (unsigned int sampleFrame = 0; sampleFrame < this->getSampleFrameCount(); sampleFrame++) {
-		for (int channel = 0; channel < _soundChannels; channel++) {
+		for (unsigned int channel = 0; channel < _soundChannels; channel++) {
 			unsigned int index = (sampleFrame * _soundChannels) + channel;
 
 			if (abs(_soundData.at(index)) >= tolerance) {
@@ -392,7 +392,7 @@ bool CX_SoundBuffer::deleteChannel(unsigned int channel) {
 	std::vector<float> dataCopy;
 
 	for (unsigned int sampleFrame = 0; sampleFrame < this->getSampleFrameCount(); sampleFrame++) {
-		for (int ch = 0; ch < this->getChannelCount(); ch++) {
+		for (unsigned int ch = 0; ch < this->getChannelCount(); ch++) {
 			if (ch != channel) {
 				unsigned int index = (sampleFrame * this->getChannelCount()) + ch;
 				dataCopy.push_back( this->_soundData[index] );
@@ -407,28 +407,28 @@ bool CX_SoundBuffer::deleteChannel(unsigned int channel) {
 
 /*!
 Sets the number of channels of the sound. Depending on the old number of channels (O) and the new number of channels (N),
-the conversion is performed in different ways. The cases in this list are evaluated in order an only 1 is executed, so a 
+the conversion is performed in different ways. The cases in this list are evaluated in order an only 1 is executed, so a
 later case cannot be reached if an earlier case has already evaluated to true. When a case says anything about the average
 of existing data, it of course means the average on a sample-by-sample basis, not the average of all the samples.
 
 + If `O == N`, nothing happens.
 + If `O == 0`, the number of channels is just set to N. However, `O == 0`, that usually means that there is no sound data available, so changing the number of channels is kind of meaningless.
 + If `N == 0`, the CX_SoundBuffer is cleared: all data is deleted. If you have no channels, you cannot have data in those channels.
-+ If `O == 1`, each of the `N` new channels is set equal to the value of the single old channel. 
++ If `O == 1`, each of the `N` new channels is set equal to the value of the single old channel.
 + If `N == 1`, and `average == true` the new channel is set equal to the average of the `O` old channels. If `average == false`, the `O - N` old channels are simply removed.
 + If `N > O`, the first `O` channels are preserved unchanged. If `average == true`, the `N - O` new channels are set to the average of the `O` old channels.
 If `average == false`, the `N - O` new channels are set to 0.
-+ If `N < O`, and `average == false`, the data from the `O - N` to-be-removed channels is discarded. 
-If `average == true` the data from the `O - N` to-be-removed channels are averaged and added on to the `N` remaining channels. 
-The averaging is done in an unusual way, so that the average intensitity of the kept channels is equal to the average intensity of the removed channels. 
++ If `N < O`, and `average == false`, the data from the `O - N` to-be-removed channels is discarded.
+If `average == true` the data from the `O - N` to-be-removed channels are averaged and added on to the `N` remaining channels.
+The averaging is done in an unusual way, so that the average intensitity of the kept channels is equal to the average intensity of the removed channels.
 An example to show why this is done:
-Assume that you have 3 channels -- a, b, and c -- and are switching to 2 channels, removing c. The average of c is just c, so when c is added to a and b, you now have c in 2 channels, 
+Assume that you have 3 channels -- a, b, and c -- and are switching to 2 channels, removing c. The average of c is just c, so when c is added to a and b, you now have c in 2 channels,
 whereas it was just in 1 channel originally: (a + c) + (b + c) = a + b + 2c. Thus, the final intensity of c is too high.
 What we want to do is scale c down by the number of channels it is being added to so that the total amount of c is equal
 both before and after changing the number of channels, so you divide c by the number of channels it is being added to (2).
 Now, (a + c/2) + (b + c/2) = a + b + c.
-However, there is another problem, which is that abs(a + c/2) can be greater than 1 even if the absolute value of both is no greater than 1. 
-Now we need to scale each sample so that it is constrained to the proper range. 
+However, there is another problem, which is that abs(a + c/2) can be greater than 1 even if the absolute value of both is no greater than 1.
+Now we need to scale each sample so that it is constrained to the proper range.
 We do that by multiplying by the number of kept channels (2) by the original number of channels (3).
 Now we have 2/3 * (a + c/2) = 2a/3 + c/3, which is bounded between -1 and 1, as long as a and c are both bounded.
 Also, 2/3 * [(a + c/2) + (b + c/2)] = 2a/3 + 2b/3 + 2c/3, so the ratios of the components of the original sound are equal.
@@ -451,7 +451,7 @@ bool CX_SoundBuffer::setChannelCount (unsigned int newChannelCount, bool average
 	}
 
 	if (O == 0) {
-		//If there are no old channels, just set the channel count to the new value. There can be no data to copy to new channels, 
+		//If there are no old channels, just set the channel count to the new value. There can be no data to copy to new channels,
 		//because there were no channels to contain the data.
 		_soundChannels = newChannelCount;
 		return true;
@@ -468,7 +468,7 @@ bool CX_SoundBuffer::setChannelCount (unsigned int newChannelCount, bool average
 
 		unsigned int originalSize = _soundData.size();
 		_soundData.resize( _soundData.size() * newChannelCount );
-		
+
 		for (unsigned int samp = 0; samp < originalSize; samp++) {
 			for (unsigned int ch = 0; ch < newChannelCount; ch++) {
 				unsigned int destIndex = _soundData.size() - 1 - (samp * newChannelCount) - ch;
@@ -480,7 +480,7 @@ bool CX_SoundBuffer::setChannelCount (unsigned int newChannelCount, bool average
 		_soundChannels = newChannelCount;
 		return true;
 	}
-	
+
 	if (N == 1) {
 		std::vector<float> newSoundData(this->getSampleFrameCount());
 
@@ -545,13 +545,13 @@ bool CX_SoundBuffer::setChannelCount (unsigned int newChannelCount, bool average
 
 		return true;
 	}
-	
+
 	if (N < O) {
 		std::vector<float> newSoundData( this->getSampleFrameCount() * newChannelCount );
 
 		if (average) {
 			//the data from the `O - N` to-be-removed channels are averaged and added on to the `N` remaining channels
-			
+
 			float removed = float(O - N);
 			//Scaling factors
 			float sigma = (float)N / (N + removed);
@@ -588,7 +588,7 @@ bool CX_SoundBuffer::setChannelCount (unsigned int newChannelCount, bool average
 	}
 	/*
 	CX::Instances::Log.error("CX_SoundBuffer") << "Sound cannot be set to the given number of channels. There is no known conversion from " <<
-						_soundChannels << " channels to " << newChannelCount << 
+						_soundChannels << " channels to " << newChannelCount <<
 						" channels. You will have to do it manually. Use getRawDataReference() to access the sound data." << endl;
 
 	return false;
@@ -598,20 +598,20 @@ bool CX_SoundBuffer::setChannelCount (unsigned int newChannelCount, bool average
 
 /*!
 Resamples the audio data stored in the CX_SoundBuffer by linear interpolation. Linear interpolation is not the ideal
-way to resample audio data; some audio fidelity is lost, more so than with other resampling techinques. It is, however, 
-very fast compared to higher-quality methods both in terms of run time and programming time. It has acceptable results, 
+way to resample audio data; some audio fidelity is lost, more so than with other resampling techinques. It is, however,
+very fast compared to higher-quality methods both in terms of run time and programming time. It has acceptable results,
 at least when the new sample rate is similar to the old sample rate.
 
 \param newSampleRate The requested sample rate.
 */
 void CX_SoundBuffer::resample (float newSampleRate) {
-	
+
 	uint64_t oldSampleCount = getSampleFrameCount();
 	uint64_t newSampleCount = (uint64_t)(getSampleFrameCount() * ((double)newSampleRate / _soundSampleRate));
 
 	vector<float> completeNewData((unsigned int)newSampleCount * _soundChannels);
 
-	for (int channel = 0; channel < _soundChannels; channel++) {
+	for (unsigned int channel = 0; channel < _soundChannels; channel++) {
 
 		for (unsigned int sample = 0; sample < newSampleCount; sample++) {
 			double time = ((double)sample)/newSampleCount;
@@ -689,7 +689,7 @@ Apply gain to the sound. The original value is simply multiplied by the amount a
 */
 bool CX_SoundBuffer::multiplyAmplitudeBy (float amount, int channel) {
 
-	if (channel >= _soundChannels) {
+	if (channel >= (int)_soundChannels) {
 		return false;
 	}
 
@@ -777,7 +777,7 @@ bool CX_SoundBuffer::writeToFile(std::string filename) {
 #define WRITE_BUFF_SIZE 4096
 
 	short writeBuff[WRITE_BUFF_SIZE];
-	int pos = 0;
+	unsigned int pos = 0;
 	while (pos < bufferSize) {
 		int len = MIN(WRITE_BUFF_SIZE, bufferSize - pos);
 		for (int i = 0; i < len; i++) {
